Remove duplicate code (#101)
* Initial refactor proposal (#93)


* Remove forensics interface (#100)


* Remove duplicate code
