+++ tests/providers/gcp/gcp_e2e.py
@@ -15,15 +15,14 @@
 """End to end test for the gcp module."""
 
 import unittest
-import json
 import time
-import os
 
 from googleapiclient.errors import HttpError
 
 from libcloudforensics.providers.gcp.internal.common import LOGGER
 from libcloudforensics.providers.gcp.internal import project as gcp_project
 from libcloudforensics.providers.gcp import forensics
+from tests.scripts import utils
 
 
 class EndToEndTest(unittest.TestCase):
@@ -54,7 +53,7 @@ class EndToEndTest(unittest.TestCase):
   @classmethod
   def setUpClass(cls):
     try:
-      project_info = ReadProjectInfo()
+      project_info = utils.ReadProjectInfo(['project_id', 'instance', 'zone'])
     except (OSError, RuntimeError, ValueError) as exception:
       raise unittest.SkipTest(str(exception))
     cls.project_id = project_info['project_id']
@@ -237,46 +236,6 @@ class EndToEndTest(unittest.TestCase):
           time.sleep(10)
 
       LOGGER.info('Disk {0:s} successfully deleted.'.format(disk))
-
-
-def ReadProjectInfo():
-  """Read project information to run e2e test.
-
-  Returns:
-    dict: A dict with the project information.
-
-  Raises:
-    OSError: If the file cannot be found, opened or closed.
-    RuntimeError: If the json file cannot be parsed.
-    ValueError: If the json file does not have the required properties.
-  """
-  project_info = os.environ.get('PROJECT_INFO')
-  if project_info is None:
-    raise OSError(
-        'Please make sure that you defined the '
-        '"PROJECT_INFO" environment variable pointing '
-        'to your project settings.')
-  try:
-    json_file = open(project_info)
-    try:
-      project_info = json.load(json_file)
-    except ValueError as exception:
-      raise RuntimeError(
-          'Cannot parse JSON file. {0:s}'.format(str(exception)))
-    json_file.close()
-  except OSError as exception:
-    raise OSError(
-        'Could not open/close file {0:s}: {1:s}'.format(
-            project_info, str(exception)))
-
-  if not all(key in project_info for key in ['project_id', 'instance', 'zone']):
-    raise ValueError(
-        'Please make sure that your JSON file '
-        'has the required entries. The file should '
-        'contain at least the following: ["project_id", '
-        '"instance", "zone"].')
-
-  return project_info
 
 
 if __name__ == '__main__':
