+++ tests/scripts/startup.sh
